+++ src/main/java/com/sandymandy/pleasurecraft/networking/S2C/PlayAttackAnimationS2CPacket.java
@@ -1,0 +1,25 @@
+package com.sandymandy.pleasurecraft.networking.S2C;
+
+import com.sandymandy.pleasurecraft.PleasureCraft;
+import net.minecraft.network.RegistryByteBuf;
+import net.minecraft.network.codec.PacketCodec;
+import net.minecraft.network.codec.PacketCodecs;
+import net.minecraft.network.packet.CustomPayload;
+import net.minecraft.util.Identifier;
+
+public record PlayAttackAnimationS2CPacket(int entityId) implements CustomPayload {
+
+    public static final Id<PlayAttackAnimationS2CPacket> ID =
+            new Id<>(Identifier.of(PleasureCraft.MOD_ID, "play_attack_animation"));
+
+    public static final PacketCodec<RegistryByteBuf, PlayAttackAnimationS2CPacket> CODEC =
+            PacketCodec.tuple(
+                PacketCodecs.VAR_INT, PlayAttackAnimationS2CPacket::entityId,
+                PlayAttackAnimationS2CPacket::new
+            );
+
+    @Override
+    public Id<? extends CustomPayload> getId() {
+        return ID;
+    }
+}